Silence integer size conversion warnings in JavaHL under Win64 by casting the values explicitly.  We assume that argument counts and property sizes are all well below the 2G limit.
* subversion/bindings/javahl/native/CopySources.cpp
  (CopySources::array): cast size_t -> int
* subversion/bindings/javahl/native/CreateJ.cpp
  (CreateJ::Checksum, 
   CreateJ::PropertyMap): ditto
* subversion/bindings/javahl/native/StringArray.cpp
  (StringArray::array): ditto
* subversion/bindings/javahl/native/SVNClient.cpp
  (SVNClient::logMessages,
   SVNClient::merge,
   SVNClient::propertyGet,
   SVNClient::revProperty): same here ...
* subversion/bindings/javahl/native/Targets.cpp
  (Targets::array): ... and here

git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1405517 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/bindings/javahl/native/CopySources.cpp
@@ -95,7 +95,7 @@ CopySources::array(SVN::Pool &pool)
   std::vector<jobject> sources = m_copySources.vector();
 
   apr_array_header_t *copySources =
-    apr_array_make(p, sources.size(), sizeof(svn_client_copy_source_t *));
+    apr_array_make(p, (int)sources.size(), sizeof(svn_client_copy_source_t *));
   for (std::vector<jobject>::const_iterator it = sources.begin();
         it < sources.end(); ++it)
     {
